rm lapVecROutines.c
