+++ src/eigenSolver.c
@@ -40,7 +40,7 @@
 #include "eigenSolver.h"
 #include "tools.h" 
 #include "linearSolver.h" // Lanczos
-#include "lapVecRoutines.h"
+#include "lapVec.h"
 #include "hamiltonianVecRoutines.h"
 #include "occupation.h"
 #include "isddft.h"
