+++ src/mixing.c
@@ -26,7 +26,7 @@
 
 #include "tools.h" 
 #include "mixing.h"
-#include "lapVecRoutines.h"
+#include "lapVec.h"
 #include "isddft.h"
 #include "linearSolver.h"
