+++ src/xc/mgga/MGGAexchangeCorrelation.c
@@ -24,7 +24,6 @@
 #include "gradVecRoutinesKpt.h"
 #include "electronicGroundState.h"
 #include "exchangeCorrelation.h"
-#include "lapVecRoutines.h"
 #include "energyDensity.h"
